requirements.txt
streamlit
numpypip show streamlit
scikit-learn
#local packages
-e .